+++ homebrew/maccrab.rb
@@ -1,5 +1,5 @@
 cask "maccrab" do
-  version "1.6.20"
+  version "1.6.21"
   sha256 "d40756aa62e2f2ecb09b27f3266e0eb01699e4fa2ac82f86fa092227c63ceb42"
 
   url "https://github.com/peterhanily/maccrab/releases/download/v#{version}/MacCrab-v#{version}.dmg"